# docs/swagger.yaml
definitions:
  handlers.APIResponse:
    properties:
      error:
        type: string
      message:
        type: string
    type: object
  models.CardMajor:
    properties:
      deck:
        example: 1
        type: integer
      id:
        type: integer
      image:
        description: Full URL
        type: string
      meanings:
        items:
          $ref: '#/definitions/models.MeaningRef'
        type: array
      name:
        example: The Fool
        type: string
      number:
        example: 0
        type: integer
      orgname:
        example: Le Mat
        type: string
    type: object
  models.CardMajorInput:
    properties:
      deck:
        example: 1
        type: integer
      name:
        example: The Fool
        type: string
      number:
        example: 0
        type: integer
      orgname:
        example: Le Mat
        type: string
    type: object
...
  models.Deck:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      sources:
        items:
          $ref: '#/definitions/models.Source'
        type: array
    type: object
  models.DeckInput:
    properties:
      description:
        type: string
      name:
        type: string
      sources:
        items:
          $ref: '#/definitions/models.IDOnly'
        type: array
    type: object
  models.DeckListItem:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.DeckRef:
    properties:
      description:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  models.IDOnly:
    properties:
      id:
        example: 1
        type: integer
    type: object
...
  models.MeaningRef:
    properties:
      id:
        type: integer
      position:
        type: string
      source:
        type: integer
    type: object
...
  models.Source:
    properties:
      decks:
        items:
          $ref: '#/definitions/models.DeckRef'
        type: array
      id:
        type: integer
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /cards/major:
    get:
      consumes:
      - application/json
      description: Returns all major arcana cards from the specified deck
      parameters:
      - description: Deck ID (required)
        in: query
        name: deckId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.CardMajor'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: List Major Arcana cards
      tags:
      - cards
    post:
      consumes:
      - application/json
      description: Adds a new card of interpretations
      parameters:
      - description: CardMajor data
        in: body
        name: card
        required: true
        schema:
          $ref: '#/definitions/models.CardMajorInput'
      produces:
      - application/json
      responses:
        "201":
          description: Returns the ID of the created card
          schema:
            $ref: '#/definitions/models.IDOnly'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Create a new card
      tags:
      - cards
  /cards/major/{id}:
    delete:
      description: Deletes a card by ID
      parameters:
      - description: CardMajor ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Delete a card
      tags:
      - cards
    get:
      description: Retrieves a card by its ID
      parameters:
      - description: CardMajor ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CardMajor'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Get card by ID
      tags:
      - cards
    put:
      consumes:
      - application/json
      description: Updates an existing card
      parameters:
      - description: CardMajor ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated card
        in: body
        name: card
        required: true
        schema:
          $ref: '#/definitions/models.CardMajorInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.CardMajor'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Update a card
      tags:
      - cards
...
  /decks:
    get:
      description: Retrieves a list of available Tarot decks. Optionally filters decks
        that contain cards.
      parameters:
      - description: Filter to include only decks that contain cards
        enum:
        - "true"
        - "yes"
        - "1"
        - "false"
        - "no"
        - "0"
        in: query
        name: hasCards
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.DeckListItem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all decks
      tags:
      - decks
    post:
      consumes:
      - application/json
      description: Adds a new Tarot deck and links to existing sources (by ID)
      parameters:
      - description: Deck input
        in: body
        name: deck
        required: true
        schema:
          $ref: '#/definitions/models.DeckInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.IDOnly'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Create a new deck
      tags:
      - decks
  /decks/{id}:
    delete:
      description: Deletes a deck and its associations
      parameters:
      - description: Deck ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Delete deck by ID
      tags:
      - decks
    get:
      description: Retrieves a deck with its associated interpretation sources
      parameters:
      - description: Deck ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Deck'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get deck by ID
      tags:
      - decks
    put:
      consumes:
      - application/json
      description: Updates the specified Tarot deck and its sources
      parameters:
      - description: Deck ID
        in: path
        name: id
        required: true
        type: integer
      - description: Deck input
        in: body
        name: deck
        required: true
        schema:
          $ref: '#/definitions/models.DeckInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Deck'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handlers.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/handlers.APIResponse'
      summary: Update a deck
      tags:
      - decks
...
swagger: "2.0"